Add global node_id to GHObject + GHTeam extends GHObject

+++ src/main/java/org/kohsuke/github/GHOrganization.java
@@ -139,7 +139,7 @@ public class GHOrganization extends GHPerson {
      *
      * @see <a href= "https://developer.github.com/v3/teams/#get-team-by-name">documentation</a>
      */
-    public GHTeam getTeam(int teamId) throws IOException {
+    public GHTeam getTeam(long teamId) throws IOException {
         return root.createRequest()
                 .withUrlPath(String.format("/organizations/%d/team/%d", id, teamId))
                 .fetch(GHTeam.class)
